# design/openapi.yaml
...
openapi: 3.0.3
info:
  title: DriveDeal — Reverse Car Marketplace API
  version: 1.0.0
  description: |-
    Backend API for DriveDeal: a reverse car marketplace where buyers post the
    car they want and dealers compete on price for it.

    Handles:
    - Buyer flows: post a car request, watch the live quote leaderboard, accept a
      quote, track the deal, chat with the dealer.
    - Dealer flows: account with manual approval, geo-filtered request feed, quote
      submission and revision, deal status.
    - Support flows: ticket queue, dealer/agent verification approval, roster.
    - AI flows (buyer only): a streaming advisory agent and a compare agent,
      grounded in the database plus a Crawl4AI web search.

    Conventions:
    - All JSON routes are under /api/v1 (this server URL). Exceptions: GET /health
      (raw health object) and the /api/v1/ws WebSocket.
    - **Success responses are envelope-free.** The payload is the resource itself.
      There is no {"success": true, "data": ...} wrapper — the frontend's mock
      client already returns bare objects and that shape is the contract
      (solution-frontend.md §2.1, solution-backend.md §0.2).
    - **Every non-2xx JSON response uses the error envelope**, and nothing else:
        {"error": {"code", "message", "details", "request_id"}}
      `code` is a stable machine-readable string the client branches on.
      `message` is for humans and may be reworded; never parse it.
...
  contact:
...
servers:
...
security:
- bearer_auth: []
paths:
...
  /cars:
    post:
      summary: Create a car listing
      description: |-
        Dealer-only, and only for an approved dealer. `status` starts at
        `draft` — a listing must be explicitly activated before it appears in
        search or in an AI source.
      operationId: createCar
      tags:
      - Inventory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CarUpsertRequest'
            examples:
              inventory_item:
                value:
                  brand_id: 1a2b3c4d-5e6f-4a7b-8c9d-0e1f2a3b4c5d
                  state_id: 0e1f2a3b-4c5d-4e6f-8a9b-0c1d2e3f4a5b
                  title: 2022 Ford Bronco Outer Banks
                  model: Bronco
                  trim: Outer Banks
                  body_type: SUV
                  seating_capacity: 5
                  condition: Used
                  mileage: 24500
                  model_year: 2022
                  fuel: Petrol
                  transmission: Automatic
                  drivetrain: 4WD
                  price: '47500.00'
                  color: '#2d6a4f'
                  color_name: Forest Green
                  description: One owner, clean CarFax, 12 months factory warranty.
                  images:
                  - https://cdn.drivedeal.example/cars/bronco-1.jpg
      responses:
        '201':
          description: Listing created as draft
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Car'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '422':
          $ref: '#/components/responses/validation_error'
    get:
      summary: List my car listings
      operationId: listMyCars
      tags:
      - Inventory
      parameters:
      - name: status
        in: query
        schema:
          $ref: '#/components/schemas/CarStatus'
      - $ref: '#/components/parameters/Cursor'
      - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Page of listings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CarList'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
  /cars/{id}:
    get:
      summary: Get a car listing
      operationId: getCar
      tags:
      - Inventory
      parameters:
      - $ref: '#/components/parameters/CarId'
      responses:
        '200':
          description: Listing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Car'
        '401':
          $ref: '#/components/responses/unauthorized'
        '404':
          $ref: '#/components/responses/not_found'
    patch:
      summary: Update a car listing
      operationId: updateCar
      tags:
      - Inventory
      parameters:
      - $ref: '#/components/parameters/CarId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CarUpsertRequest'
      responses:
        '200':
          description: Listing updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Car'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '404':
          $ref: '#/components/responses/not_found'
        '422':
          $ref: '#/components/responses/validation_error'
  /cars/{id}/status:
    post:
      summary: Change a listing's status
      description: |-
        `draft → active → sold` / `removed`, server-validated. Only an `active`
        listing may be attached to a quote and cited as an AI source.
      operationId: updateCarStatus
      tags:
      - Inventory
      parameters:
      - $ref: '#/components/parameters/CarId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - status
              properties:
                status:
                  $ref: '#/components/schemas/CarStatus'
      responses:
        '200':
          description: Status changed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Car'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '404':
          $ref: '#/components/responses/not_found'
        '409':
          $ref: '#/components/responses/conflict'
...
components:
  parameters:
...
    CarId:
      name: id
      in: path
      required: true
      description: Car listing UUID
      schema:
        type: string
        format: uuid
...
    Cursor:
      name: cursor
      in: query
      description: Opaque pagination cursor from the previous page
      schema:
        type: string
    Limit:
      name: limit
      in: query
      description: Page size
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  securitySchemes:
    bearer_auth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |-
        RS256 access token from /auth/login or /auth/refresh.

        Subject = `profiles.id` (NOT the `users.id`, which is a different
        UUID and does not appear anywhere in the request layer). Roles are
        read from the `profiles` row on every request, so a role change or a
        suspension takes effect on the next call rather than at token expiry.

        For the WebSocket, use a one-shot ticket from /chats/ws-ticket
        instead — a browser cannot set this header on the handshake.
  responses:
    unauthorized:
      description: |-
        Missing, malformed, expired, or revoked token. `code`
        distinguishes: `UNAUTHORIZED`, `TOKEN_EXPIRED`, `TOKEN_REVOKED`.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            expired:
              value:
                error:
                  code: TOKEN_EXPIRED
                  message: Your session has expired. Sign in again.
                  details: null
                  request_id: 3f2a1b0c-9d8e-4f7a-8b6c-5d4e3f2a1b0c
    forbidden:
      description: |-
        Authenticated, but not allowed. `code` distinguishes:
        `FORBIDDEN_ROLE` (wrong role for this surface), `NOT_RESOURCE_OWNER`
        (someone else's resource), `INSUFFICIENT_ROLE` (a support agent trying
        to approve a peer agent).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            wrong_role:
              value:
                error:
                  code: FORBIDDEN_ROLE
                  message: The AI advisor is only available to buyers.
                  details: null
                  request_id: 7a6b5c4d-3e2f-4a1b-8c9d-0e1f2a3b4c5d
            agent_approving_agent:
              value:
                error:
                  code: INSUFFICIENT_ROLE
                  message: Only an admin can approve a support agent account.
                  details:
                    required_role: admin
                    your_role: support
                  request_id: 1b2c3d4e-5f60-4a1b-8c2d-3e4f5a6b7c8d
...
    not_found:
      description: The resource does not exist, or the caller may not know it exists
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            missing:
              value:
                error:
                  code: NOT_FOUND
                  message: That request could not be found.
                  details: null
                  request_id: 4d5e6f70-8a9b-4c1d-9e2f-3a4b5c6d7e8f
    conflict:
      description: |-
        The request is valid but the current state forbids it. `code`
        distinguishes: `CONFLICT` (generic, `details` lists the legal
        transitions), `ALREADY_EXISTS` (duplicate signup, second live quote
        on the same request), `INVALID_TRANSITION` (deal or quote status
        machine rejected the move), `TOKEN_ALREADY_USED` (refresh replay).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            illegal_deal_transition:
              value:
                error:
                  code: INVALID_TRANSITION
                  message: A matched deal cannot move straight to completed.
                  details:
                    current: matched
                    attempted: completed
                    allowed_transitions:
                    - negotiation
                    - cancelled
                  request_id: 5e6f7081-9a2b-4c3d-8e4f-5a6b7c8d9e0f
            duplicate_quote:
              value:
                error:
                  code: ALREADY_EXISTS
                  message: You already have a live quote on this request. Revise it instead.
                  details:
                    existing_quote_id: 724c2d61-b1be-47ba-9458-0f4001204a1b
                  request_id: 6f708192-0a3b-4d5e-9f6a-7b8c9d0e1f2a
    validation_error:
      description: |-
        Body, path or query failed validation. `details` is a
        field → message map so the client can highlight the offending input
        rather than showing a generic banner.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            field_errors:
              value:
                error:
                  code: VALIDATION_ERROR
                  message: Some fields need attention.
                  details:
                    email:
                    - That email is already registered.
                    budget_max:
                    - budget_max must be greater than or equal to budget_min.
                  request_id: 0718293a-4b5c-4d6e-8f70-9a1b2c3d4e5f
...
  schemas:
    DecimalString:
      type: string
      description: |-
        A decimal serialised as a string, e.g. "70229.00". **Never a JSON
        number** — JavaScript would turn 0.1 + 0.2 into 0.30000000000000004
        and an OTD estimate would be off by cents.

        Other observed values: `47500.00`, `0.065`.
      pattern: ^-?\d+(\.\d{1,2})?$
      example: '70229.00'
    ErrorResponse:
      type: object
      description: |-
        The only non-2xx JSON shape. There is no `success: false` wrapper —
        the presence of `error` is the signal.
      required:
      - error
      properties:
        error:
          type: object
          required:
          - code
          - message
          properties:
            code:
              type: string
              description: |-
                Stable machine-readable code. Branch on this, never on
                `message`.

                All codes: UNAUTHORIZED, TOKEN_EXPIRED, FORBIDDEN_ROLE, NOT_RESOURCE_OWNER, INSUFFICIENT_ROLE, DEALER_PENDING_REVIEW, SUPPORT_PENDING_REVIEW, NOT_FOUND, CONFLICT, ALREADY_EXISTS, INVALID_TRANSITION, VALIDATION_ERROR, RATE_LIMITED, INTERNAL_ERROR.
              example: UNAUTHORIZED
            message:
              type: string
              description: Human-readable. May be reworded at any time — do not parse.
              example: The AI advisor is only available to buyers.
            details:
              description: |-
                Structured context for this code — a field→message map for
                VALIDATION_ERROR, legal transitions for INVALID_TRANSITION,
                allowed_transitions for CONFLICT, retry_after for
                RATE_LIMITED. Null when there is nothing useful to add.
              type: object
              additionalProperties: true
              nullable: true
            request_id:
              type: string
              description: Correlates with the server log line and the `X-Request-ID` response header.
              example: 3f2a1b0c-9d8e-4f7a-8b6c-5d4e3f2a1b0c
...
    BodyType:
      type: string
      description: |-
        Taxonomy-driven and expected to grow, so this is not a hard CHECK
        constraint in the database. Clients must tolerate an unknown value
        and render it rather than crash.

        Observed values: SUV, Sedan, Hatchback, Sports Car, Coupe, MUV, Luxury, Truck, Van, Wagon, Convertible, Supercar, Crossover.
      example: SUV
    CarCondition:
      type: string
      enum:
      - New
      - Used
      - Certified
    Transmission:
      type: string
      enum:
      - Automatic
      - Manual
      - CVT
      - Dual Clutch
      - Other
    Drivetrain:
      type: string
      enum:
      - FWD
      - RWD
      - AWD
      - 4WD
      - Other
    FuelType:
      type: string
      enum:
      - Petrol
      - Diesel
      - Hybrid
      - Electric
      - CNG
      - LPG
      - Other
...
    CarStatus:
      type: string
      enum:
      - draft
      - active
      - sold
      - removed
      description: Only `active` listings may be quoted and cited as an AI source.
...
    CarUpsertRequest:
      type: object
      required:
      - brand_id
      - title
      - model
      - model_year
      - price
      properties:
        brand_id:
          type: string
          format: uuid
        state_id:
          type: string
          format: uuid
          description: Required before the listing can appear in a geo-filtered feed.
          nullable: true
        title:
          type: string
          minLength: 1
          maxLength: 200
          description: Denormalised display label, e.g. "2022 Ford Bronco Outer Banks".
        model:
          type: string
          minLength: 1
          maxLength: 120
        trim:
          type: string
          maxLength: 120
          nullable: true
        body_type:
          allOf:
          - $ref: '#/components/schemas/BodyType'
          nullable: true
        seating_capacity:
          type: integer
          minimum: 1
          maximum: 15
          description: |-
            A fact about this specific car, and what a buyer's
            `seater_count` is matched against. Newer than the live CSVs, so
            it is null on migrated rows until the dealer or the advisor
            pipeline fills it — and null means "no constraint", not "zero
            seats".
          nullable: true
        condition:
          allOf:
          - $ref: '#/components/schemas/CarCondition'
          nullable: true
        mileage:
          type: integer
          minimum: 0
          nullable: true
        model_year:
          type: integer
          minimum: 1990
          maximum: 2035
        fuel:
          allOf:
          - $ref: '#/components/schemas/FuelType'
          nullable: true
        transmission:
          allOf:
          - $ref: '#/components/schemas/Transmission'
          nullable: true
        drivetrain:
          allOf:
          - $ref: '#/components/schemas/Drivetrain'
          nullable: true
        price:
          $ref: '#/components/schemas/DecimalString'
        color:
          type: string
          description: Hex, e.g. "#2d6a4f"
          pattern: ^#[0-9a-fA-F]{6}$
          nullable: true
        color_name:
          type: string
          description: What buyers actually filter on, via `buyer_preference.exterior_color`.
          nullable: true
        description:
          type: string
          maxLength: 5000
          nullable: true
        reviews:
          type: array
          description: |-
            Review objects, the shape below. Kept in a `jsonb` array on the
            listing — there is deliberately **no separate `car_reviews`
            table**, and the 15-table budget has no room for one.
          items:
            $ref: '#/components/schemas/CarReview'
        images:
          type: array
          description: CDN or S3 URLs. Empty array, never null.
          items:
            type: string
            format: uri
    CarReview:
      type: object
      required:
      - rating
      - body
      - source
      - created_at
      properties:
        rating:
          type: number
          minimum: 0
          maximum: 5
          example: 4.5
        title:
          type: string
          example: Great car, minor niggles
          nullable: true
        body:
          type: string
          example: Owned for 18 months. Engine is smooth, but infotainment lags.
        author_name:
          type: string
          nullable: true
        source:
          type: string
          description: |-
            Where the review came from. `owner_reported` for a verified
            owner, `crawled` for one Crawl4AI pulled from a public page —
            which is how reviews get written back into the database now that
            embeddings are gone.

            Values: `owner_reported`, `crawled`, `dealer`.
          example: owner_reported
        created_at:
          type: string
          format: date-time
    Car:
      allOf:
      - type: object
        required:
        - id
        - seller_id
        - brand_id
        - title
        - model
        - model_year
        - price
        - status
        - created_at
        - updated_at
        properties:
          id:
            type: string
            format: uuid
          seller_id:
            type: string
            format: uuid
            description: |-
              The owning **dealer's** profile UUID. The column is still
              named `seller_id` for historical reasons, but the role is
              `dealer` everywhere in the API.
          brand_id:
            type: string
            format: uuid
          brand_name:
            type: string
            nullable: true
          state_id:
            type: string
            format: uuid
            nullable: true
          title:
            type: string
          model:
            type: string
          trim:
            type: string
            nullable: true
          body_type:
            allOf:
            - $ref: '#/components/schemas/BodyType'
            nullable: true
          seating_capacity:
            type: integer
            nullable: true
          condition:
            allOf:
            - $ref: '#/components/schemas/CarCondition'
            nullable: true
          mileage:
            type: integer
            nullable: true
          model_year:
            type: integer
          fuel:
            allOf:
            - $ref: '#/components/schemas/FuelType'
            nullable: true
          transmission:
            allOf:
            - $ref: '#/components/schemas/Transmission'
            nullable: true
          drivetrain:
            allOf:
            - $ref: '#/components/schemas/Drivetrain'
            nullable: true
          price:
            $ref: '#/components/schemas/DecimalString'
          color:
            type: string
            nullable: true
          color_name:
            type: string
            nullable: true
          description:
            type: string
            nullable: true
          rating:
            type: number
            minimum: 0
            maximum: 5
            description: Cached average of the `reviews` array.
            nullable: true
          reviews:
            type: array
            items:
              $ref: '#/components/schemas/CarReview'
          images:
            type: array
            items:
              type: string
              format: uri
          status:
            $ref: '#/components/schemas/CarStatus'
          views:
            type: integer
            minimum: 0
          otd:
            allOf:
            - $ref: '#/components/schemas/DecimalString'
            description: |-
              `price * (1 + tax_rate)` for the listing's state, computed
              server-side. Sent so the buyer never does money arithmetic
              in the client, and the figure a dealer quotes is compared
              against is the same one everywhere.
          created_at:
            type: string
            format: date-time
          updated_at:
            type: string
            format: date-time
    CarList:
      type: object
      required:
      - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Car'
        next_cursor:
          type: string
          nullable: true
...